requirements.txt
# local package
-e .

# external requirements
click
alabaster
alembic
appdirs
APScheduler
atpublic
attrs
autopep8
awscli
Babel
botocore
cached-property
cachetools
certifi
cffi
chardet
cloudpickle
colorama
colorhash
commonmark
configobj
configparser
coverage
cryptography
cycler
databricks-cli
decorator
dictdiffer
diskcache
distlib
distro
docker
docutils
dpath
dulwich
dvc
entrypoints
filelock
flake8
Flask
Flask-Cors
Flask-MonitoringDashboard
flatten-dict
flufl.lock
fsspec
ftfy
funcy
future
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
grandalf
gunicorn
httplib2
idna
imagesize
importlib-metadata
iniconfig
itsdangerous
Jinja2
jmespath
joblib
jsonpath-ng
kiwisolver
kneed
livereload
lunr
mailchecker
Mako
Markdown
MarkupSafe
matplotlib
mccabe
mkdocs
mkdocs-material
mkdocs-material-extensions
mlflow
nanotime
networkx
nltk
numpy
oauth2client
packaging
pandas
pathlib2
pathspec
phonenumbers
Pillow
pluggy
ply
prometheus-client
prometheus-flask-exporter
protobuf
psutil
py
pyasn1
pyasn1-modules
pycodestyle
pycparser
pydantic
pydot
PyDrive2
pyflakes
pygit2
Pygments
pygtrie
pymdown-extensions
pyOpenSSL
pyparsing
pytest
python-benedict
python-dateutil
python-dotenv
python-editor
python-fsutil
python-slugify
pytz
PyYAML
querystring-parser
regex
requests
rich
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
scikit-learn
scipy
shortuuid
shtab
six
smmap
snowballstemmer
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
SQLAlchemy
sqlparse
tabulate
text-unidecode
threadpoolctl
toml
tornado
tox
tqdm
typing-extensions
tzlocal
uritemplate
urllib3
virtualenv
voluptuous
wcwidth
websocket-client
Werkzeug
xmltodict
zc.lockfile
zipp